Ejercicio 2.2

	new file:   .gitignore
	new file:   Makefile
	copied:     main.c -> gol.c
	new file:   gol.h
	modified:   main.c
